Add Metal views for iOS and tvOS Remove whitespaces Move iOS and tvOS ViewController to Window class file Remake OS X OpenGL initialization Call sharedEngine->begin() in applicationDidFinishLaunching on OS X, iOS, tvOS

+++ ouzel/DrawNode.cpp
@@ -191,7 +191,7 @@ namespace ouzel
             std::vector<uint16_t> indices;
             std::vector<video::VertexPC> vertices;
 
-            for(uint32_t i = 0; i < 3; ++i)
+            for(uint16_t i = 0; i < 3; ++i)
             {
                 indices.push_back(i);
